+++ crates/pluot_core/src/layers/filled_curve_layer.rs
@@ -2,8 +2,7 @@
 // and ultimately renders a TriangulatedLayer as a sub-layer.
 // This layer is intended to be used as a sub-layer of CurveLayer.
 
-use earcut::Earcut;
-use kurbo::{Arc as KurboArc, CubicBez, ParamCurve, Point as KurboPoint, QuadBez, SvgArc, Vec2 as KurboVec2};
+use kurbo::{CubicBez, ParamCurve};
 use serde::{Deserialize, Serialize};
 use std::sync::Arc;
 
@@ -17,170 +16,10 @@ use crate::two::shapes::{TwoColor, TwoElement, TwoGroup, TwoPath};
 use crate::two::svg::{update_svg, SvgContext};
 use crate::wgpu;
 
-use super::curve_and_polygon_utils::resolve_margins;
+use super::curve_and_polygon_utils::{
+    commands_to_subpaths, compute_fill_vertices, resolve_margins, PathCommand,
+};
 use super::triangulated_layer::{TriangulatedLayer, TriangulatedLayerParams};
-
-// ── PathCommand ────────────────────────────────────────────────────────────────
-
-/// A single drawing command, the post-parsed form of an SVG path segment.
-/// All coordinates are absolute.
-#[derive(Serialize, Deserialize, Debug, Clone)]
-#[serde(tag = "type", rename_all = "snake_case")]
-pub enum PathCommand {
-    MoveTo { x: f32, y: f32 },
-    LineTo { x: f32, y: f32 },
-    CubicTo { x1: f32, y1: f32, x2: f32, y2: f32, x: f32, y: f32 },
-    QuadraticTo { x1: f32, y1: f32, x: f32, y: f32 },
-    ArcTo {
-        rx: f32,
-        ry: f32,
-        #[serde(default)]
-        x_axis_rotation: f32,
-        large_arc: bool,
-        sweep: bool,
-        x: f32,
-        y: f32,
-    },
-    Close,
-}
-
-// ── Curve helpers ──────────────────────────────────────────────────────────────
-
-fn line_to_cubic(p0: KurboPoint, p1: KurboPoint) -> CubicBez {
-    CubicBez::new(p0, p0, p1, p1)
-}
-
-pub(crate) fn commands_to_subpaths(commands: &[PathCommand]) -> Vec<Vec<CubicBez>> {
-    let mut subpaths: Vec<Vec<CubicBez>> = Vec::new();
-    let mut current: Vec<CubicBez> = Vec::new();
-    let mut cursor = KurboPoint::ZERO;
-    let mut subpath_start = KurboPoint::ZERO;
-
-    for command in commands {
-        match *command {
-            PathCommand::MoveTo { x, y } => {
-                if !current.is_empty() {
-                    subpaths.push(std::mem::take(&mut current));
-                }
-                cursor = KurboPoint::new(x as f64, y as f64);
-                subpath_start = cursor;
-            }
-            PathCommand::LineTo { x, y } => {
-                let end = KurboPoint::new(x as f64, y as f64);
-                current.push(line_to_cubic(cursor, end));
-                cursor = end;
-            }
-            PathCommand::CubicTo { x1, y1, x2, y2, x, y } => {
-                let end = KurboPoint::new(x as f64, y as f64);
-                current.push(CubicBez::new(
-                    cursor,
-                    KurboPoint::new(x1 as f64, y1 as f64),
-                    KurboPoint::new(x2 as f64, y2 as f64),
-                    end,
-                ));
-                cursor = end;
-            }
-            PathCommand::QuadraticTo { x1, y1, x, y } => {
-                let end = KurboPoint::new(x as f64, y as f64);
-                let quad = QuadBez::new(cursor, KurboPoint::new(x1 as f64, y1 as f64), end);
-                current.push(quad.raise());
-                cursor = end;
-            }
-            PathCommand::ArcTo { rx, ry, x_axis_rotation, large_arc, sweep, x, y } => {
-                let end = KurboPoint::new(x as f64, y as f64);
-                let svg_arc = SvgArc {
-                    from: cursor,
-                    to: end,
-                    radii: KurboVec2::new(rx as f64, ry as f64),
-                    x_rotation: (x_axis_rotation as f64).to_radians(),
-                    large_arc,
-                    sweep,
-                };
-                match KurboArc::from_svg_arc(&svg_arc) {
-                    Some(arc) => {
-                        let tolerance = (svg_arc.radii.hypot() * 1e-3).max(1e-9);
-                        let mut p0 = cursor;
-                        arc.to_cubic_beziers(tolerance, |p1, p2, p3| {
-                            current.push(CubicBez::new(p0, p1, p2, p3));
-                            p0 = p3;
-                        });
-                    }
-                    None => current.push(line_to_cubic(cursor, end)),
-                }
-                cursor = end;
-            }
-            PathCommand::Close => {
-                if cursor != subpath_start {
-                    current.push(line_to_cubic(cursor, subpath_start));
-                }
-                cursor = subpath_start;
-            }
-        }
-    }
-    if !current.is_empty() {
-        subpaths.push(current);
-    }
-    subpaths
-}
-
-fn subpath_to_ring(subpath: &[CubicBez], subdivisions: u32) -> Vec<(f32, f32)> {
-    let mut points: Vec<(f32, f32)> = Vec::new();
-    if subpath.is_empty() {
-        return points;
-    }
-    let push_unique = |points: &mut Vec<(f32, f32)>, p: (f32, f32)| {
-        match points.last() {
-            Some(last) if (last.0 - p.0).abs() < 1e-9 && (last.1 - p.1).abs() < 1e-9 => {}
-            _ => points.push(p),
-        }
-    };
-    let start = subpath[0].p0;
-    push_unique(&mut points, (start.x as f32, start.y as f32));
-    for seg in subpath {
-        for step in 1..=subdivisions {
-            let t = step as f64 / subdivisions as f64;
-            let p = seg.eval(t);
-            push_unique(&mut points, (p.x as f32, p.y as f32));
-        }
-    }
-    if points.len() > 1 {
-        let first = points[0];
-        let last = *points.last().unwrap();
-        if (first.0 - last.0).abs() < 1e-9 && (first.1 - last.1).abs() < 1e-9 {
-            points.pop();
-        }
-    }
-    points
-}
-
-fn ring_area_2x(points: &[(f32, f32)]) -> f32 {
-    let n = points.len();
-    let mut area = 0.0f32;
-    for i in 0..n {
-        let j = (i + 1) % n;
-        area += points[i].0 * points[j].1 - points[j].0 * points[i].1;
-    }
-    area
-}
-
-fn compute_fill_vertices(subpaths: &[Vec<CubicBez>], subdivisions: u32) -> Vec<(f32, f32)> {
-    let mut verts: Vec<(f32, f32)> = Vec::new();
-    let mut ec: Earcut<f32> = Earcut::new();
-    let mut indices: Vec<u32> = Vec::new();
-    for subpath in subpaths {
-        let ring = subpath_to_ring(subpath, subdivisions);
-        if ring.len() < 3 || ring_area_2x(&ring).abs() <= 1e-12 {
-            continue;
-        }
-        ec.earcut(ring.iter().map(|&(x, y)| [x, y]), &[] as &[u32], &mut indices);
-        for &i in &indices {
-            verts.push(ring[i as usize]);
-        }
-    }
-    verts
-}
-
-// ── Params ─────────────────────────────────────────────────────────────────────
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 #[serde(default)]
@@ -214,8 +53,6 @@ impl Default for FilledCurveLayerParams {
     }
 }
 
-// ── Layer ──────────────────────────────────────────────────────────────────────
-
 pub struct FilledCurveLayer {
     view_params: ViewParams,
     layer_params: FilledCurveLayerParams,
@@ -225,6 +62,8 @@ pub struct FilledCurveLayer {
 
 impl FilledCurveLayer {
     pub fn new(view_params: ViewParams, layer_params: FilledCurveLayerParams) -> Self {
+        // TODO: move the triangulation into the prepare() function?
+        // TODO: only do the triangulation in the raster drawing case?
         let subpaths = commands_to_subpaths(&layer_params.commands);
         let subdivisions = layer_params.subdivisions.max(1);
         let fill_vertices = Arc::new(compute_fill_vertices(&subpaths, subdivisions));
@@ -232,15 +71,11 @@ impl FilledCurveLayer {
     }
 }
 
-// ── Helpers ────────────────────────────────────────────────────────────────────
-
-
-// ── Trait impls ────────────────────────────────────────────────────────────────
-
 #[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
 #[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
 impl PreparedLayer for FilledCurveLayer {
     async fn prepare(&mut self, _gpu_context: Option<&GpuContext<'_>>) -> PrepareResult {
+        // TODO: run the TriangulatedLayer sub-layer's prepare function here?
         PrepareResult { bailed_early: false }
     }
 }
